Remove global caches
As documented by the cachetools project, Cache objects are NOT
thread-safe and need to be explicitly synchronized when shared.

Remove global cache objects and make them instance members of the Client
classes. This shifts responsibility for ensuring thread-safety to the
calling code.

This has the obvious side-effect of preventing multiple Client objects
from sharing the same cache.